Basic operations completed
After a lot of struggle understanding how to implement SOAP with Spring Boot, I finally made it work! However, a lot of work still has to be done: business-logic validations (including server-to-server communication with the UserService), and appointment retrievals, but I think the hardest part is completed.

+++ AppointmentService/src/main/java/dtos/BookingDTO.java
@@ -13,7 +13,7 @@ public class BookingDTO {
     
     private AppointmentDTO appointment;
     
-    private Integer referenceNumber;
+    private Long referenceNumber;
     
     private BookingStatus status;
 
@@ -23,7 +23,7 @@ public class BookingDTO {
             Integer id, 
             Integer patientId, 
             AppointmentDTO appointment, 
-            Integer referenceNumber, 
+            Long referenceNumber, 
             BookingStatus status
     ) {
         this.id = id;
@@ -45,9 +45,9 @@ public class BookingDTO {
 
     public void setAppointment(AppointmentDTO appointment) {this.appointment = appointment;}
 
-    public Integer getReferenceNumber() {return referenceNumber;}
+    public Long getReferenceNumber() {return referenceNumber;}
 
-    public void setReferenceNumber(Integer referenceNumber) {this.referenceNumber = referenceNumber;}
+    public void setReferenceNumber(Long referenceNumber) {this.referenceNumber = referenceNumber;}
 
     public BookingStatus getStatus() {return status;}
